file: db: metrics call
lots of refactoring/optimizations here.

we avoid the stdlib os package because it's super allocation heavy
and slow. indeed, we even skip the dirent parsing code in syscall
and reimplement it to avoid allocations!

the performance benefits are vast: we reduce allocations from
about 2GB to 2MB in the 100k metrics case, and the time by a
factor of 10.